moved exceptions from training
