+++ tests/integration_tests/detection_dataset_test.py
@@ -14,7 +14,7 @@ from super_gradients.training.dataloaders.dataloaders import _process_dataset_pa
 from super_gradients.training.datasets import PascalVOCDetectionDataset, COCODetectionDataset
 from super_gradients.training.transforms import DetectionMosaic, DetectionPaddedRescale, DetectionTargetsFormatTransform
 from super_gradients.training.datasets.data_formats.default_formats import XYXY_LABEL
-from super_gradients.training.exceptions.dataset_exceptions import EmptyDatasetException
+from super_gradients.common.exceptions.dataset_exceptions import EmptyDatasetException
 from super_gradients.common.environment.path_utils import normalize_path
